lucuhelp/requirements.txt
Django==1.8.2
PyJWT==1.3.0
httplib2==0.9.1
psycopg2==2.6.1
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
wsgiref==0.1.2
celery==3.1.18
django-celery==3.1.16